[ADD]: Animated background

+++ source/utils/Vec3.cpp
@@ -4,7 +4,7 @@ Vec3::Vec3(void) {
 	x = y = z = 0;
 }
 
-Vec3::Vec3(int x, int y, int z) {
+Vec3::Vec3(float x, float y, float z) {
 	this->x = x;
 	this->y = y;
 	this->z = z;
